requirements.txt
PyAutoGUI==0.9.53
Pillow==9.2.0

